Add switch for readonly watchlists
Bug: T160062

+++ includes/specials/SpecialEditWatchlist.php
@@ -451,6 +451,10 @@ class SpecialEditWatchlist extends UnlistedSpecialPage {
 	 * Remove all titles from a user's watchlist
 	 */
 	private function clearWatchlist() {
+		if ( $this->getConfig()->get( 'ReadOnlyWatchedItemStore' ) ) {
+			throw new DBReadOnlyError( null, 'The watchlist is currently readonly.' );
+		}
+
 		$dbw = wfGetDB( DB_MASTER );
 		$dbw->delete(
 			'watchlist',
